Refactor QuickSearch into DataGridSearch
